Included diagrams (new formats), added dev schedule. [WIP]
Diagrams are added in with vector graphics in eps format. These were
created using Inkscape's pdf to esp conversion. The eps file types are
the ones included in the latex document, and on execution of tex to pdf
these images are also converted to pdf vector images to be included in
the final document.