Fixed secret.go bug, complete interpolation.go - closes #18, closes #19
There was a bug in secret.go when constructing a polynomial for a
certain threshold. For a threshold of k, a polynomial of degree k-1
should have been constructed, but instead a polynomial of k degrees was
being created. Noticed this error during testing when a secret shared
among two people could not be reconstructed.

Completed the interpolation.go functions, clearing up issues with bad
interpolation due to an integer division instead of multiplication by a
modular inverse.

Also go fmt'd the src directory.