Added election generator; creates voters, election key, shares...
Includes additions to the configuration for a blockchain. The generator
will create a json file usable by the blockchain package. It will
contain a voter's private DSA key, a voter's vote token, a list of all
vote tokens and their corresponding public DSA keys, a list of peers to
connect to, the public Election key, and the voter's secret share of the
private Election key. Note there are two shares, one for each component
of the private Election key, along with two modulus values which are
public knowledge. The modulus values are required for the reconstruction
of the private Election key. Added also is a boolean value SyncPeers
which currently has no effect, but will be used to allow a peer to not
automatically sync their peer list every X seconds. This serves the
purpose of creating a type of network where every node is not connected
to every other node.